\name{getNet}
\alias{getNet}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Gets a network structure
}
\description{
This function is simply a wrapper around the plotting method for mlVAR objects, that extracts the network structure rather than plotting them.
}
\usage{
getNet(x, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
An `mlVAR` or `mlVARsim0` object.
}
  \item{\dots}{
Arguments sent to \code{\link{plot.mlVAR}}
}
}
\author{
Sacha Epskamp <mail@sachaepskamp.com>
}

