\name{importMplus}
\alias{importMplus}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Import output from Mplus
}
\description{
This function imports the output from an Mplus model that has been generated by mlVAR. It can be used to make manual changes to the input file. 
}
\usage{
importMplus(outfile)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{outfile}{
Location of Mplus output file.
}
}

\author{
Sacha Epskamp <mail@sachaepskamp.com>
}

