\name{plot.psyvar}
\alias{plot.psyvar}

\title{
  Plot method for "psyvar"
}
\description{
  Plots an object created by \code{\link{psyvar}}.
}
\usage{
  \method{plot}{qgraph}(object, type = c("fixed","se","random","subject"), lag = 1,subject,...)
}

\arguments{
  \item{object}{
    A "psyvar" object
  }
  \item{type}{
  
  
  }
  \item{lag}{
  
  
  }
  \item{subject}{
  
  }
  \item{\dots}{
    Arguments sent to \code{\link[qgraph]{qgraph}}
  }
}

\details{
 
}

\author{
  Sacha Epskamp (mail@sachaepskamp.com)
}
