\name{mlVAR-methods}
\alias{print.mlVAR}
\alias{summary.mlVAR}

\title{
  print and summary functions for mlVAR objects
}
\description{
  Create a short summary of an object created by \code{\link{mlVAR}}.
}
\usage{
  \method{print}{mlVAR}(x, ...)
  \method{summary}{mlVAR}(object, ...)
}

\arguments{
  \item{object}{
    A "mlVAR" object
  }
    \item{x}{
    A "mlVAR" object
  }
    \item{\dots}{
    Not used
  }
}

\author{
  Sacha Epskamp (mail@sachaepskamp.com), Marie K. Deserno (m.k.deserno@uva.nl) and Laura F. Bringmann (laura.bringmann@ppw.kuleuven.be)
}
