\name{mlVAR0-methods}
\alias{print.mlVAR0}
\alias{summary.mlVAR0}

\title{
  print and summary functions for mlVAR0 objects
}
\description{
  Create a short summary of an object created by \code{\link{mlVAR0}}.
  
  FUNCTION IS DEPRECATED AND WILL BE REMOVED SOON.
}
\usage{
  \method{print}{mlVAR0}(x, ...)
  \method{summary}{mlVAR0}(object, ...)
}

\arguments{
  \item{object}{
    A "mlVAR0" object
  }
    \item{x}{
    A "mlVAR0" object
  }
    \item{\dots}{
    Not used
  }
}

\author{
  Sacha Epskamp (mail@sachaepskamp.com), Marie K. Deserno (m.k.deserno@uva.nl) and Laura F. Bringmann (laura.bringmann@ppw.kuleuven.be)
}
