\name{mlVAR-effects}
\alias{fixedEffects}
\alias{randomEffects}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Fixed and random effects
}
\description{
These functions return a table of the fixed and random effects.

FUNCTIONS ARE DEPRECATED AND WILL BE REMOVED SOON.
}
\usage{
fixedEffects(object, digits = 5)
randomEffects(object, digits = 5)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{object}{
A \code{mlVAR} object
}
 \item{digits}{
    Number of digits to output
 }
}


\author{
Sacha Epskamp (mail@sachaepskamp.com), Marie K. Deserno (m.k.deserno@uva.nl) and Laura F. Bringmann (laura.bringmann@ppw.kuleuven.be)
}
